pyproject.toml
[tool.poetry]
name = "digital_library_django_ninja"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
pytest = "^7.1.2"
python-dotenv = "^0.20.0"
django-ninja = "^0.19.1"
psycopg = {extras = ["binary", "pool"], version = "^3.0.16"}

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
